Make irc an optional dependency
