+++ .github/workflows/rust.yml
@@ -71,6 +71,10 @@ jobs:
           - name: default
           - name: none
             flags: --no-default-features
+          - name: git
+            flags: --no-default-features -F git
+          - name: irc
+            flags: --no-default-features -F irc
     steps:
     - uses: actions/checkout@v4
